+++ packages/terra-core-docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added test example for controlled form native select component.
 
 * Changed
   * Updated `terra-status-view - Change Variant` example component to use the apply button to change the selected variant.